bot/requirements.txt
aiogram>=3.0.0
motor>=3.1.0
aiohttp>=3.8.0
python-dateutil>=2.8.2
pymongo>=4.3.0
ujson>=5.4.0
python-dotenv>=0.21.0
mlab_amo_async==0.0.2
